Get data from GXPA

Get data from GXPA for a given series id and genes

Usage

get_data_from_gxpa(series_id, genes, GXPA_TOKEN = Sys.getenv("GXPA_TOKEN"))

Arguments

series_id

dataset series id

genes

select genes

GXPA_TOKEN

GXPA API token. Default is to use the environment variable GXPA_TOKEN

Value

a data.frame of expression values for selected genes and sample annotation information.

Details

Go to GXPA Settings to get your token. Some datasets require you to login, so be sure to do that before you copy your token.

Token can be set by:

For the current session:

Sys.setenv(GXPA_TOKEN = "XXXXXXXXXXXXXXXX")

For all sessions:

usethis::edit_r_environ(),

adding: GXPA_TOKEN = "XXXXXXXXXXXXXXXX"

Examples

dat <- get_data_from_gxpa(series_id = 41, genes = c("CD33", "CD34"))
#> These genes not present:
#> CD33
#> CD34
head(dat)
#>            Cd34 Cd33      cell_type
#> GSM2107347 9.48 5.22   hypofunc_CD8
#> GSM2107348 9.10 5.22   hypofunc_CD8
#> GSM2107349 9.55 5.49   hypofunc_CD8
#> GSM2107350 5.95 5.29 functional_CD8
#> GSM2107351 5.91 5.29 functional_CD8
#> GSM2107352 6.01 5.36 functional_CD8
#>                                                                                                                                                                                                                                                                                                                                                  description
#> GSM2107347                         Gene expression data from hypofunctional tumor (AH1)-specific CD8+ T cells taken from an approximately 1 cm long CT26 tumor (15 days of growth) after therapeutic vaccination with mimotope A5 (11 days post vaccination). Details regarding mimotope A5 have been published at Slansky et al. 2000 & Jordan et al. 2010.
#> GSM2107348                         Gene expression data from hypofunctional tumor (AH1)-specific CD8+ T cells taken from an approximately 1 cm long CT26 tumor (15 days of growth) after therapeutic vaccination with mimotope A5 (11 days post vaccination). Details regarding mimotope A5 have been published at Slansky et al. 2000 & Jordan et al. 2010.
#> GSM2107349                         Gene expression data from hypofunctional tumor (AH1)-specific CD8+ T cells taken from an approximately 1 cm long CT26 tumor (15 days of growth) after therapeutic vaccination with mimotope A5 (11 days post vaccination). Details regarding mimotope A5 have been published at Slansky et al. 2000 & Jordan et al. 2010.
#> GSM2107350 Gene expression data from functional tumor (AH1)-specific CD8+ T cells taken from the spleen of an approximately 1 cm long CT26 tumor-bearing mouse (15 days of growth) after therapeutic vaccination with mimotope A5 (11 days post vaccination). Details regarding mimotope A5 have been published at Slansky et al. 2000 & Jordan et al. 2010.
#> GSM2107351 Gene expression data from functional tumor (AH1)-specific CD8+ T cells taken from the spleen of an approximately 1 cm long CT26 tumor-bearing mouse (15 days of growth) after therapeutic vaccination with mimotope A5 (11 days post vaccination). Details regarding mimotope A5 have been published at Slansky et al. 2000 & Jordan et al. 2010.
#> GSM2107352 Gene expression data from functional tumor (AH1)-specific CD8+ T cells taken from the spleen of an approximately 1 cm long CT26 tumor-bearing mouse (15 days of growth) after therapeutic vaccination with mimotope A5 (11 days post vaccination). Details regarding mimotope A5 have been published at Slansky et al. 2000 & Jordan et al. 2010.
#>                     title     tissue
#> GSM2107347        TIL1+A5 CT26_tumor
#> GSM2107348        TIL1+A5 CT26_tumor
#> GSM2107349        TIL1+A5 CT26_tumor
#> GSM2107350 Spleen_TIL1+A5     spleen
#> GSM2107351 Spleen_TIL1+A5     spleen
#> GSM2107352 Spleen_TIL1+A5     spleen
#>                                                             source_name_ch1
#> GSM2107347                             CT26 tumor + therapeutic vaccination
#> GSM2107348                             CT26 tumor + therapeutic vaccination
#> GSM2107349                             CT26 tumor + therapeutic vaccination
#> GSM2107350 Spleen from mouse bearing a CT26 tumor + therapeutic vaccination
#> GSM2107351 Spleen from mouse bearing a CT26 tumor + therapeutic vaccination
#> GSM2107352 Spleen from mouse bearing a CT26 tumor + therapeutic vaccination