Get Series Information from GXPA

Get series information from GXPA for a all data or a given series id

Usage

get_series_info_from_gxpa(
  series_id = NULL,
  GXPA_TOKEN = Sys.getenv("GXPA_TOKEN")
)

Arguments

series_id

dataset series id(s). NULL (default) will output all dataset series information

GXPA_TOKEN

GXPA API token. Default is to use the environment variable GXPA_TOKEN

Value

a data.frame of series information for selected datasets, with the following columns:

id: series ID

name: dataset name ID

description: dataset description

organism: dataset organism

perm_id: dataset perm_id

Details

Go to GXPA Settings to get your token. Some datasets require you to login, so be sure to do that before you copy your token.

Token can be set by:

For the current session:

Sys.setenv(GXPA_TOKEN = "XXXXXXXXXXXXXXXX")

For all sessions:

usethis::edit_r_environ(),

adding: GXPA_TOKEN = "XXXXXXXXXXXXXXXX"

Examples

get_series_info_from_gxpa(41)
#>    id                                             name
#> 11 41 GSE79858 antigen specific hypofunctional T cells
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 description
#> 11 Molecular profile of tumor-specific CD8+ T cell hypofunction in cancer; Comparison of genome-wide mRNA expresson between tumor-infiltrating CD8+ T cells from the tumor (hypofunctional T cells) and periphery (functional T cells); Mechanisms of self-tolerance often result in CD8+ tumor-infiltrating lymphocytes (TIL) with a hypofunctional phenotype incapable of tumor clearance. Using a solid tumor model in mice, we found that CD8+ T cells became tolerized in less than 24 hours in an established tumor environment. To define the collective impact of pathways suppressing TIL function, we compared genome-wide mRNA expression of tumor-specific CD8+ T cells from the tumor and periphery. Notably, gene expression induced during TIL hypofunction more closely resembled self-tolerance than viral-exhaustion. Differential gene expression was refined to identify a core set of genes that defined hypofunctional TIL; these data comprise the first “molecular profile” of tumor-specific TIL that are naturally responding and represent a polyclonal repertoire. The molecular profile of TIL was further dissected to determine the extent of overlap and distinction between pathways that collectively restrict T cell functions. As suggested by the molecular profile of TIL, protein expression of inhibitory receptor LAG-3 was differentially regulated throughout prolonged late-G1/early-S phase of the cell cycle. Our data may accelerate efficient identification of combination therapies to boost anti-tumor function of TIL specifically against tumor cells.
#>    organism perm_id
#> 11    mouse       1
# Can give a numeric vector
get_series_info_from_gxpa(41:43)
#>    id                                             name
#> 11 41 GSE79858 antigen specific hypofunctional T cells
#> 12 43           GSE86881 LCMV antigen specific T cells
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 description
#> 11 Molecular profile of tumor-specific CD8+ T cell hypofunction in cancer; Comparison of genome-wide mRNA expresson between tumor-infiltrating CD8+ T cells from the tumor (hypofunctional T cells) and periphery (functional T cells); Mechanisms of self-tolerance often result in CD8+ tumor-infiltrating lymphocytes (TIL) with a hypofunctional phenotype incapable of tumor clearance. Using a solid tumor model in mice, we found that CD8+ T cells became tolerized in less than 24 hours in an established tumor environment. To define the collective impact of pathways suppressing TIL function, we compared genome-wide mRNA expression of tumor-specific CD8+ T cells from the tumor and periphery. Notably, gene expression induced during TIL hypofunction more closely resembled self-tolerance than viral-exhaustion. Differential gene expression was refined to identify a core set of genes that defined hypofunctional TIL; these data comprise the first “molecular profile” of tumor-specific TIL that are naturally responding and represent a polyclonal repertoire. The molecular profile of TIL was further dissected to determine the extent of overlap and distinction between pathways that collectively restrict T cell functions. As suggested by the molecular profile of TIL, protein expression of inhibitory receptor LAG-3 was differentially regulated throughout prolonged late-G1/early-S phase of the cell cycle. Our data may accelerate efficient identification of combination therapies to boost anti-tumor function of TIL specifically against tumor cells.
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Transcriptional profiles of virus-specific CD8+ T cells from control or anti-PD-L1-treated LCMV-Clone 13 infected mice one day or 18-29 weeks after treatment; RNA-Seq analysis was used to study the profile of exhausted Tcells and how anti-PD-L1 treatment changed the properties of these cells short term & long term after treatment
#>    organism perm_id
#> 11    mouse       1
#> 12    mouse       1

# Can get all available data
all_dat <- get_series_info_from_gxpa()
all_dat$name[1:10]
#>  [1] "GSE49910 : Immune Cells - Tom Freeman"    
#>  [2] "GSE17946 : Tcells with various stims"     
#>  [3] "GSE7307: Body Atlas, R Roth"              
#>  [4] "GSE28490 : immune cells, Hoffmann-LaRoche"
#>  [5] "GSE47552 PCs from disease"                
#>  [6] "GSE61597"                                 
#>  [7] "GSE68891: B and PC from MM patients"      
#>  [8] "GSE36769 CD4 TILs"                        
#>  [9] "GSE12453"                                 
#> [10] "GSE85947 PD1 TIM3 mouse TILs"             
nrow(all_dat)
#> [1] 139
table(all_dat$organism)
#> 
#> human mouse 
#>   113    26 

# Can get series ID based on dataset name
all_dat$id[all_dat$name == "Beat AML - OHSU"]
#> [1] 151