Load session data to database

Run the remote function to upload data from the session to the app database. If this session has not been loaded to the database, then load_type must be upload_new which is the default. On the other hand, if this session is already matched to one in the database, then several other choices for load_type are available: update_series add_new_scores and load_coords. For more information on these options, go to the Load Data page on the app.

A message will be shown to say if the task submission was successful or error message if there was a problem.

Usage

load_session(
  session_id,
  load_type = c("upload_new", "update_series", "add_new_scores", "load_coords"),
  user_cookie = login_and_get_user_cookie(),
  server_url = Sys.getenv("GXPA_SERVER", "https://geneatlas.redda.bms.com/")
)

Arguments

session_id

the id of session

load_type

Default 'upload_new'. See description above for more choices.

user_cookie

the cookie string. Can use login_and_get_user_cookie() to get it. If you don't provide the cookie string, then the function will run login_and_get_user_cookie(). So, if you provide the cookie then it saves one interaction with the server.

server_url

GXPA server (default GXPA_SERVER environment variable or https://geneatlas.redda.bms.com/ if missing)

Value

Invisibly returns the full html output which can be useful in case you didn't get a success message.

Examples

if (FALSE) { # \dontrun{
# This session (X1NBJK) on the demo app has fake data and is tiny and easy to test
load_session("X1NBJK", load_type = "upload_new")
load_session("X1NBJK", load_type = "load_scores")
} # }