Create Series Info File for GXPA

Create Series Info File for GXPA

Usage

make_series_info_file(
  filename,
  series_name,
  series_descript,
  default_group,
  organism,
  expr_units,
  pmid = NULL,
  analyst = NULL,
  analyst_comments = NULL,
  default_color = NULL,
  extra.list = NULL
)

Arguments

filename

series info output file name

series_name

series name (can match series name used in GXPA app)

series_descript

series description

default_group

default group to consider for plots (will often be used for x axis variable)

organism

series organism

expr_units

data experimental units

pmid

PubMed id (can be NULL)

analyst

data analyst (can be NULL)

analyst_comments

comments by analyst (can be NULL)

default_color

default color to consider for plots (can be NULL)

extra.list

named list of extra variables to include in the series info file

Value

writes out csv file of series information to filename location

Examples

filename <- tempfile(fileext = ".csv")
make_series_info_file(
  file = filename,
  series_name = "SU2C_2019",
  series_descript = "Long description",
  default_group = "tumor_site",
  organism = "human",
  expr_units = "RNA seq log2 (FPKM+1)",
  pmid = 31061129,
  default_color = "AR_score_group",
  extra.list = list(
    "Extra_Field" = "Something",
    "Another_one" = 5
  )
)
utils::read.csv(filename)
#>               key                 value
#> 1     series_name             SU2C_2019
#> 2 series_descript      Long description
#> 3   default_group            tumor_site
#> 4        organism                 human
#> 5      expr_units RNA seq log2 (FPKM+1)
#> 6            pmid              31061129
#> 7   default_color        AR_score_group
#> 8     Extra_Field             Something
#> 9     Another_one                     5