Remove session directory

Remove a session directory. This is permanent. It will fail if the session is linked to a series. To unlink a session and a series, you need to use the web interface to delete the series. Then you can remove the session directory.

Usage

remove_session(
  session_id,
  user_cookie = login_and_get_user_cookie(),
  server_url = Sys.getenv("GXPA_SERVER", "https://geneatlas.redda.bms.com/")
)

Arguments

session_id

the id of session

user_cookie

the cookie string. Can use login_and_get_user_cookie() to get it. If you don't provide the cookie string, then the function will run login_and_get_user_cookie(). So, if you provide the cookie then it saves one interaction with the server.

server_url

GXPA server (default GXPA_SERVER environment variable or https://geneatlas.redda.bms.com/ if missing)

Value

Invisibly returns the full html output which can be useful in case you didn't get a success message.

See also

Examples

if (FALSE) { # \dontrun{
user_cookie = login_and_get_user_cookie()
sess_id = begin_new_session("new_test_session", user_cookie=user_cookie)
remove_session(sess_id)
} # }