Push a File to a GXPA Session

Importing data for an already opened GXPA session.

Usage

send_file_to_session(
  session_id,
  filename_local,
  file_save_name,
  user_cookie = login_and_get_user_cookie(),
  server_url = Sys.getenv("GXPA_SERVER", "https://geneatlas.redda.bms.com/")
)

Arguments

session_id

session id from a session that has already been opened in GXPA

filename_local

input file name, with path, to import

file_save_name

file name to be saved as in GXPA. Must match a know name (see details)

user_cookie

user cookie (default is using login_and_get_user_cookie())

server_url

GXPA server (default GXPA_SERVER environment variable or https://geneatlas.redda.bms.com/ if missing)

Value

message indicating upload was successful

Details

The GXPA session needs to exist, so go to the GXPA app to see open sessions and create a new one if needed. The session id can be found on the main sessions page or in the URL once in the session (6 uppercase chars or numbers)

Common file_save_name values are expr.expr.txt, expr.samples.csv, expr.series_info.csv, and expr.score_info.csv, for expression data loading, and GEO.expr.txt, GEO.samples.csv, GEO.series_info.csv, and GEO.score_info.csv, for GEO data loading

Examples

if (FALSE) { # \dontrun{
tmp_file <- tempfile()
utils::write.csv(matrix(rnorm(100), nrow = 10), tmp_file)
send_file_to_session(
  session_id = "7AR22L",
  filename_local = tmp_file,
  file_save_name = "expr.expr.txt"
)
} # }